+++ src/main/java/module/teamAnalyzer/ui/FilterPanel.java
@@ -1,5 +1,6 @@
 package module.teamAnalyzer.ui;
 
+import core.file.xml.TeamStats;
 import core.gui.HOMainFrame;
 import core.gui.comp.panel.ImagePanel;
 import core.model.HOVerwaltung;
@@ -8,27 +9,33 @@ import module.teamAnalyzer.SystemManager;
 import module.teamAnalyzer.ht.HattrickManager;
 import module.teamAnalyzer.manager.TeamManager;
 import module.teamAnalyzer.vo.Team;
+import org.jetbrains.annotations.NotNull;
 
 import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
+import java.util.Map;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
 import javax.swing.*;
 
 /**
- * Panel to filter opponents matches.
+ * Panel to filter and download opponent's matches.
  */
-public class FilterPanel extends JPanel implements ActionListener {
+public class FilterPanel extends JPanel {
 
 	private static final String CARD_AUTOMATIC = "AUTOMATIC CARD";
 	private static final String CARD_MANUAL = "MANUAL CARD";
 	private static boolean teamComboUpdating = false;
 	private AutoFilterPanel autoPanel;
 	private final JButton downloadButton = new JButton(HOVerwaltung.instance().getLanguageString("ls.button.update"));
+	private final JButton analyzeButton = new JButton(HOVerwaltung.instance().getLanguageString("AutoFilterPanel.Analyze"));
 	private final JComboBox<Team> teamCombo = new JComboBox<>();
 	private final JPanel cards = new JPanel(new CardLayout());
 	private JRadioButton radioAutomatic;
 	private JRadioButton radioManual;
 	private ManualFilterPanel manualPanel;
+
+	private final TeamInfoPanel teamInfoPanel = new TeamInfoPanel();
 
 	/**
 	 * Creates a new FilterPanel object.
@@ -45,25 +52,6 @@ public class FilterPanel extends JPanel implements ActionListener {
 	}
 
 	/**
-	 * Handle action events.
-	 */
-	@Override
-	public void actionPerformed(ActionEvent ae) {
-		Object compo = ae != null ? ae.getSource() : null;
-		CardLayout cLayout = (CardLayout) (cards.getLayout());
-
-		if (radioAutomatic.equals(compo)) {
-			TeamAnalyzerPanel.filter.setAutomatic(true);
-			autoPanel.reload();
-			cLayout.show(cards, CARD_AUTOMATIC);
-		} else if (radioManual.equals(compo)) {
-			cLayout.show(cards, CARD_MANUAL);
-			TeamAnalyzerPanel.filter.setAutomatic(false);
-			manualPanel.reload();
-		}
-	}
-
-	/**
 	 * Update GUI elements.
 	 */
 	public void reload() {
@@ -73,6 +61,7 @@ public class FilterPanel extends JPanel implements ActionListener {
 
 		downloadButton.setEnabled(true);
 		downloadButton.setText(HOVerwaltung.instance().getLanguageString("ls.button.update"));
+		analyzeButton.setEnabled(true);
 
 		CardLayout cLayout = (CardLayout) (cards.getLayout());
 
@@ -80,8 +69,7 @@ public class FilterPanel extends JPanel implements ActionListener {
 			radioAutomatic.setSelected(true);
 			cLayout.show(cards, CARD_AUTOMATIC);
 			autoPanel.reload();
-		}
-		else {
+		} else {
 			radioManual.setSelected(true);
 			cLayout.show(cards, CARD_MANUAL);
 			manualPanel.reload();
@@ -122,13 +110,20 @@ public class FilterPanel extends JPanel implements ActionListener {
 		teamCombo.setOpaque(false);
 		teamCombo.addItemListener(e -> {
 			if (!teamComboUpdating) {
-                SystemManager.setActiveTeam((Team)teamCombo.getSelectedItem());
+				Team selectedTeam = (Team) teamCombo.getSelectedItem();
+				SystemManager.setActiveTeam(selectedTeam);
+        assert selectedTeam != null;
+        Map<String, String> teamDetails = retrieveTeamDetails(selectedTeam);
+				teamInfoPanel.setTeam(teamDetails);
 				SystemManager.refresh();
 			}
 		});
 
-		JButton analyzeButton = new JButton(HOVerwaltung.instance().getLanguageString(
-				"AutoFilterPanel.Analyze"));
+		if (teamCombo.getSelectedItem() != null) {
+			Team selectedTeam = (Team) teamCombo.getSelectedItem();
+			Map<String, String> teamDetails = retrieveTeamDetails(selectedTeam);
+			teamInfoPanel.setTeam(teamDetails);
+		}
 
 		analyzeButton.addActionListener(e -> {
 			if (radioManual.isSelected()) {
@@ -139,39 +134,74 @@ public class FilterPanel extends JPanel implements ActionListener {
 			SystemManager.updateReport();
 		});
 
-		downloadButton.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
+		downloadButton.addActionListener(e -> {
+            downloadButton.setEnabled(false);
+			// Trigger event in a separate thread to avoid Button UI from being blocked.
+			SwingUtilities.invokeLater(() -> {
+				final ExecutorService downloadExecutor = Executors.newCachedThreadPool();
+
 				HOLogger.instance().log(getClass(),
 						"UPDATE for Team " + SystemManager.getActiveTeamId());
-				//HattrickManager.downloadPlayers(SystemManager.getActiveTeamId());
+
 				// Load squad info of all teams
-				for ( var team  : TeamManager.getTeams()){
-					HattrickManager.downloadPlayers(team.getTeamId());
+				try {
+					for (var team : TeamManager.getTeams()) {
+						downloadExecutor.execute(() -> HattrickManager.downloadPlayers(team.getTeamId()));
+					}
+
+					downloadExecutor.execute(() ->
+							HattrickManager.downloadMatches(SystemManager.getActiveTeamId(), TeamAnalyzerPanel.filter));
+				} finally {
+					downloadExecutor.shutdown();
+					try {
+						downloadExecutor.awaitTermination(30, TimeUnit.SECONDS);
+						analyzeButton.setEnabled(true);
+					} catch (Exception ee) {
+						HOLogger.instance().error(FilterPanel.class, "Error awaiting termination: "  + ee.getMessage());
+					}
 				}
-				HattrickManager.downloadMatches(SystemManager.getActiveTeamId(), TeamAnalyzerPanel.filter);
 				HOMainFrame.instance().setInformationCompleted();
 				SystemManager.refresh();
-			}
-		});
+
+				HOLogger.instance().info(getClass(),
+						"Download complete for Team " + SystemManager.getActiveTeamId());
+
+				downloadButton.setEnabled(true);
+			});
+        });
+
+		JPanel mainTeamPanel = new JPanel();
+		mainTeamPanel.setLayout(new BorderLayout());
 
 		JPanel teamPanel = new ImagePanel();
 		teamPanel.setBorder(BorderFactory.createEmptyBorder(4, 4, 4, 4));
 
 		teamPanel.setLayout(new BorderLayout());
-		teamPanel.add(downloadButton, BorderLayout.NORTH);
 		teamPanel.add(teamCombo, BorderLayout.SOUTH);
 		teamPanel.setOpaque(false);
+
+		mainTeamPanel.add(teamPanel, BorderLayout.NORTH);
+		mainTeamPanel.add(teamInfoPanel, BorderLayout.CENTER);
 
 		JPanel topPanel = new ImagePanel();
 
 		topPanel.setLayout(new BorderLayout());
 		radioAutomatic = new JRadioButton(HOVerwaltung.instance().getLanguageString("Option.Auto")); //$NON-NLS-1$
 		radioAutomatic.setSelected(true);
-		radioAutomatic.addActionListener(this);
+		radioAutomatic.addActionListener(e -> {
+			final CardLayout cLayout = (CardLayout) cards.getLayout();
+			TeamAnalyzerPanel.filter.setAutomatic(true);
+			autoPanel.reload();
+			cLayout.show(cards, CARD_AUTOMATIC);
+		});
 		radioAutomatic.setOpaque(false);
 		radioManual = new JRadioButton(HOVerwaltung.instance().getLanguageString("Manual")); //$NON-NLS-1$
-		radioManual.addActionListener(this);
+		radioManual.addActionListener(e -> {
+			final CardLayout cLayout = (CardLayout) cards.getLayout();
+			cLayout.show(cards, CARD_MANUAL);
+			TeamAnalyzerPanel.filter.setAutomatic(false);
+			manualPanel.reload();
+		});
 		radioManual.setOpaque(false);
 
 		ButtonGroup groupRadio = new ButtonGroup();
@@ -187,16 +217,37 @@ public class FilterPanel extends JPanel implements ActionListener {
 		buttonPanel.add(radioManual);
 		buttonPanel.add(Box.createHorizontalGlue());
 
-		topPanel.add(teamPanel, BorderLayout.NORTH);
-		topPanel.add(buttonPanel, BorderLayout.SOUTH);
+		topPanel.add(mainTeamPanel, BorderLayout.NORTH);
 
-		main.add(topPanel, BorderLayout.NORTH);
-		add(main, BorderLayout.NORTH);
+		JPanel downloadGamesPanel = new JPanel();
+		downloadGamesPanel.setLayout(new BorderLayout());
+		downloadGamesPanel.add(buttonPanel, BorderLayout.NORTH);
+		downloadGamesPanel.setBorder(BorderFactory.createTitledBorder("Download Games"));
+
 		autoPanel = new AutoFilterPanel();
 		manualPanel = new ManualFilterPanel();
 		cards.add(autoPanel, CARD_AUTOMATIC);
 		cards.add(manualPanel, CARD_MANUAL);
-		add(cards, BorderLayout.CENTER);
-		add(analyzeButton, BorderLayout.SOUTH);
+		downloadGamesPanel.add(cards, BorderLayout.CENTER);
+
+		JPanel buttonContainerPanel = new JPanel(new GridLayout(1, 2, 4 ,4));
+		buttonContainerPanel.setOpaque(true);
+		buttonContainerPanel.add(downloadButton);
+		buttonContainerPanel.add(analyzeButton);
+		downloadGamesPanel.add(buttonContainerPanel, BorderLayout.SOUTH);
+
+		main.add(topPanel, BorderLayout.NORTH);
+		main.add(downloadGamesPanel, BorderLayout.CENTER);
+		add(main, BorderLayout.CENTER);
+	}
+
+	@NotNull
+	private static Map<String, String> retrieveTeamDetails(Team selectedTeam) {
+		Map<String, String> teamDetails = HattrickManager.getTeamDetails(selectedTeam.getTeamId());
+		TeamStats teamStats = HattrickManager.downloadSeriesDetails(Integer.parseInt(teamDetails.get("LeagueLevelUnitID")), selectedTeam.getTeamId());
+		if (teamStats != null) {
+			teamDetails.put("LeaguePosition", String.valueOf(teamStats.getPosition()));
+		}
+		return teamDetails;
 	}
 }